Move picosha2 to extern
