+++ CMakeLists.txt
@@ -94,6 +94,7 @@ target_link_libraries(Etterna PRIVATE pcre)
 target_link_libraries(Etterna PRIVATE libmad)
 target_link_libraries(Etterna PRIVATE stb)
 target_link_libraries(Etterna PRIVATE libcurl)
+target_link_libraries(Etterna PRIVATE picosha2)
 
 # OS Specific Initialization
 if(WIN32)
