+++ src/Etterna/Globals/CMakeLists.txt
@@ -19,8 +19,7 @@ list(APPEND GLOBAL_FILES_HPP
 	"ProductInfo.h" # TODO: Have this be auto-generated.
 	"SpecialFiles.h"
 	"rngthing.h"
-	"StepMania.h" # TODO: Refactor into separate main project.
-	"picosha2.h")
+	"StepMania.h") # TODO: Refactor into separate main project.
 
 if(NOT APPLE)
 	list(APPEND GLOBAL_FILES_SRC "Main.cpp")
